requirements.txt
fastapi
uvicorn
python-multipart
aiogram
aiosqlite
jinja2
python-dotenv